code clean-up, qfopen function bug fix
