+++ tests/CMakeLists.txt
@@ -11,4 +11,4 @@ endif()
 add_subdirectory(firmware)
 add_subdirectory(identify)
 add_subdirectory(modem)
-add_subdirectory(gsm_console)
+add_subdirectory(modem_http)
